Refactor TodayPage and Add today due filter
